clearing_house: market_index as u16 (#124)
* init

* smalelr market indexes

* lint/prettier fixes

+++ programs/clearing_house/src/controller/funding.rs
@@ -126,7 +126,7 @@ pub fn settle_funding_payments(
 
 #[allow(clippy::comparison_chain)]
 pub fn update_funding_rate(
-    market_index: u64,
+    market_index: u16,
     market: &mut PerpMarket,
     oracle_map: &mut OracleMap,
     now: UnixTimestamp,
